Components folder hierarchy restructured
