+++ src/ducks/Studio/Template/Button.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import cx from 'classnames'
 import Icon from '@santiment-network/ui/Icon'
-import FormDialogNewTemplate from './FormDialog/NewTemplate'
+import FormDialogNewTemplate from './Dialog/NewTemplate'
 import LoginDialog from '../../../components/LoginDialog'
 import styles from './index.module.scss'
